US7768 - Map Implementation Organize map components
