+++ src/examples/map/app.component.html
@@ -1,0 +1,1 @@
+<app-map></app-map>
