+++ src/examples/map/index.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <base href="." />
-    <title>angular2 playground</title>
+    <title>angular2-google-maps</title>
     <link rel="stylesheet" href="/app.css" />
     <script src="https://unpkg.com/zone.js/dist/zone.js"></script>
     <script src="https://unpkg.com/zone.js/dist/long-stack-trace-zone.js"></script>
