+++ src/examples/map/map-content.component.ts
@@ -1,0 +1,210 @@
+import { Component, AfterViewInit } from '@angular/core';
+
+import { GoogleMapsAPIWrapper } from 'angular2-google-maps/core';
+
+declare let google: any;
+
+interface NativeGoogMapProps {
+  zoomControlOptions?: any;
+  streetViewControlOptions?: any;
+}
+
+@Component({
+  selector: 'app-map-content',
+  template: ''
+})
+export class MapContentComponent implements AfterViewInit {
+
+  constructor(public mapApiWrapper: GoogleMapsAPIWrapper) {}
+
+  public ngAfterViewInit() {
+    this.mapApiWrapper.getNativeMap().then((map) => {
+      map.setOptions(<NativeGoogMapProps> {
+        zoomControlOptions: {
+          style: google.maps.ControlPosition.small,
+          position: google.maps.ControlPosition.LEFT_TOP
+        },
+        streetViewControlOptions: {
+          position: google.maps.ControlPosition.RIGHT_CENTER
+        },
+        minZoom: 3,
+        maxZoom: 20,
+        scrollwheel: false,
+        styles: [
+          {
+            'elementType': 'geometry',
+            'stylers': [
+              {
+                'color': '#f5f5f5'
+              }
+            ]
+          },
+          {
+            'elementType': 'labels.icon',
+            'stylers': [
+              {
+                'visibility': 'off'
+              }
+            ]
+          },
+          {
+            'elementType': 'labels.text.fill',
+            'stylers': [
+              {
+                'color': '#616161'
+              }
+            ]
+          },
+          {
+            'elementType': 'labels.text.stroke',
+            'stylers': [
+              {
+                'color': '#f5f5f5'
+              }
+            ]
+          },
+          {
+            'featureType': 'administrative.land_parcel',
+            'elementType': 'labels.text.fill',
+            'stylers': [
+              {
+                'color': '#bdbdbd'
+              }
+            ]
+          },
+          {
+            'featureType': 'administrative.province',
+            'elementType': 'geometry.stroke',
+            'stylers': [
+              {
+                'color': '#979797'
+              },
+              {
+                'weight': 1
+              }
+            ]
+          },
+          {
+            'featureType': 'poi',
+            'elementType': 'geometry',
+            'stylers': [
+              {
+                'color': '#eeeeee'
+              }
+            ]
+          },
+          {
+            'featureType': 'poi',
+            'elementType': 'labels.text.fill',
+            'stylers': [
+              {
+                'color': '#757575'
+              }
+            ]
+          },
+          {
+            'featureType': 'poi.park',
+            'elementType': 'geometry',
+            'stylers': [
+              {
+                'color': '#e5e5e5'
+              }
+            ]
+          },
+          {
+            'featureType': 'poi.park',
+            'elementType': 'labels.text.fill',
+            'stylers': [
+              {
+                'color': '#9e9e9e'
+              }
+            ]
+          },
+          {
+            'featureType': 'road',
+            'elementType': 'geometry',
+            'stylers': [
+              {
+                'color': '#ffffff'
+              }
+            ]
+          },
+          {
+            'featureType': 'road.arterial',
+            'elementType': 'labels.text.fill',
+            'stylers': [
+              {
+                'color': '#757575'
+              }
+            ]
+          },
+          {
+            'featureType': 'road.highway',
+            'elementType': 'geometry',
+            'stylers': [
+              {
+                'color': '#dadada'
+              }
+            ]
+          },
+          {
+            'featureType': 'road.highway',
+            'elementType': 'labels.text.fill',
+            'stylers': [
+              {
+                'color': '#616161'
+              }
+            ]
+          },
+          {
+            'featureType': 'road.local',
+            'elementType': 'labels.text.fill',
+            'stylers': [
+              {
+                'color': '#9e9e9e'
+              }
+            ]
+          },
+          {
+            'featureType': 'transit.line',
+            'elementType': 'geometry',
+            'stylers': [
+              {
+                'color': '#e5e5e5'
+              }
+            ]
+          },
+          {
+            'featureType': 'transit.station',
+            'elementType': 'geometry',
+            'stylers': [
+              {
+                'color': '#eeeeee'
+              }
+            ]
+          },
+          {
+            'featureType': 'water',
+            'elementType': 'geometry',
+            'stylers': [
+              {
+                'color': '#c9c9c9'
+              }
+            ]
+          },
+          {
+            'featureType': 'water',
+            'elementType': 'labels.text.fill',
+            'stylers': [
+              {
+                'color': '#9e9e9e'
+              }
+            ]
+          }
+        ],
+        fullscreenControl: false
+      });
+    });
+  }
+
+}
