+++ src/examples/map/map.component.ts
@@ -1,210 +1,46 @@
-import { Component, AfterViewInit } from '@angular/core';
-
-import { GoogleMapsAPIWrapper } from 'angular2-google-maps/core';
-
-declare let google: any;
-
-interface NativeGoogMapProps {
-  zoomControlOptions?: any;
-  streetViewControlOptions?: any;
-}
+import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-map-content',
-  template: ''
+  selector: 'app-map',
+  styles: [`
+    .sebm-google-map-container {
+       height: 300px;
+     }
+  `],
+  templateUrl: './map.component.html'
 })
-export class MapComponent implements AfterViewInit {
+export class MapComponent {
 
-  constructor(public mapApiWrapper: GoogleMapsAPIWrapper) {}
+  // google maps zoom level
+  zoom = 8;
 
-  public ngAfterViewInit() {
-    this.mapApiWrapper.getNativeMap().then((map) => {
-      map.setOptions(<NativeGoogMapProps> {
-        zoomControlOptions: {
-          style: google.maps.ControlPosition.small,
-          position: google.maps.ControlPosition.LEFT_TOP
-        },
-        streetViewControlOptions: {
-          position: google.maps.ControlPosition.RIGHT_CENTER
-        },
-        minZoom: 3,
-        maxZoom: 20,
-        scrollwheel: false,
-        styles: [
-          {
-            'elementType': 'geometry',
-            'stylers': [
-              {
-                'color': '#f5f5f5'
-              }
-            ]
-          },
-          {
-            'elementType': 'labels.icon',
-            'stylers': [
-              {
-                'visibility': 'off'
-              }
-            ]
-          },
-          {
-            'elementType': 'labels.text.fill',
-            'stylers': [
-              {
-                'color': '#616161'
-              }
-            ]
-          },
-          {
-            'elementType': 'labels.text.stroke',
-            'stylers': [
-              {
-                'color': '#f5f5f5'
-              }
-            ]
-          },
-          {
-            'featureType': 'administrative.land_parcel',
-            'elementType': 'labels.text.fill',
-            'stylers': [
-              {
-                'color': '#bdbdbd'
-              }
-            ]
-          },
-          {
-            'featureType': 'administrative.province',
-            'elementType': 'geometry.stroke',
-            'stylers': [
-              {
-                'color': '#979797'
-              },
-              {
-                'weight': 1
-              }
-            ]
-          },
-          {
-            'featureType': 'poi',
-            'elementType': 'geometry',
-            'stylers': [
-              {
-                'color': '#eeeeee'
-              }
-            ]
-          },
-          {
-            'featureType': 'poi',
-            'elementType': 'labels.text.fill',
-            'stylers': [
-              {
-                'color': '#757575'
-              }
-            ]
-          },
-          {
-            'featureType': 'poi.park',
-            'elementType': 'geometry',
-            'stylers': [
-              {
-                'color': '#e5e5e5'
-              }
-            ]
-          },
-          {
-            'featureType': 'poi.park',
-            'elementType': 'labels.text.fill',
-            'stylers': [
-              {
-                'color': '#9e9e9e'
-              }
-            ]
-          },
-          {
-            'featureType': 'road',
-            'elementType': 'geometry',
-            'stylers': [
-              {
-                'color': '#ffffff'
-              }
-            ]
-          },
-          {
-            'featureType': 'road.arterial',
-            'elementType': 'labels.text.fill',
-            'stylers': [
-              {
-                'color': '#757575'
-              }
-            ]
-          },
-          {
-            'featureType': 'road.highway',
-            'elementType': 'geometry',
-            'stylers': [
-              {
-                'color': '#dadada'
-              }
-            ]
-          },
-          {
-            'featureType': 'road.highway',
-            'elementType': 'labels.text.fill',
-            'stylers': [
-              {
-                'color': '#616161'
-              }
-            ]
-          },
-          {
-            'featureType': 'road.local',
-            'elementType': 'labels.text.fill',
-            'stylers': [
-              {
-                'color': '#9e9e9e'
-              }
-            ]
-          },
-          {
-            'featureType': 'transit.line',
-            'elementType': 'geometry',
-            'stylers': [
-              {
-                'color': '#e5e5e5'
-              }
-            ]
-          },
-          {
-            'featureType': 'transit.station',
-            'elementType': 'geometry',
-            'stylers': [
-              {
-                'color': '#eeeeee'
-              }
-            ]
-          },
-          {
-            'featureType': 'water',
-            'elementType': 'geometry',
-            'stylers': [
-              {
-                'color': '#c9c9c9'
-              }
-            ]
-          },
-          {
-            'featureType': 'water',
-            'elementType': 'labels.text.fill',
-            'stylers': [
-              {
-                'color': '#9e9e9e'
-              }
-            ]
-          }
-        ],
-        fullscreenControl: false
-      });
-    });
-  }
+  // initial center position for the map
+  lat = 39.159398;
+  lng = -84.423367;
+
+  disableDefaultUi = false;
+  zoomControl = true;
+
+  // markers: marker[] = [
+  markers = <any> [
+    {
+      lat: this.lat + 0.3,
+      lng: this.lng + 0.3,
+      label: 'A',
+      draggable: false
+    },
+    {
+      lat: this.lat - 0.3,
+      lng: this.lng - 0.3,
+      label: 'B',
+      draggable: false
+    },
+    {
+      lat: this.lat - 0.45,
+      lng: this.lng - 0.45,
+      label: 'C',
+      draggable: false
+    }
+  ];
 
 }
